backend/tests/requirements.txt
requests>=2.28.0
pytest-playwright>=0.10.0
